+++ pkg/store/store.go
@@ -1,552 +1,502 @@
 package store
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
-	"sync"
 	"time"
 
-	"github.com/adrg/xdg"
+	"github.com/arthur-debert/nanostore/nanostore"
 	"github.com/arthur-debert/padz/pkg/config"
-	"github.com/arthur-debert/padz/pkg/filelock"
 	"github.com/arthur-debert/padz/pkg/filesystem"
 	"github.com/arthur-debert/padz/pkg/logging"
 	"github.com/arthur-debert/padz/pkg/project"
 )
 
+// Constants
 const (
-	dataDirName        = "scratch"
-	metadataFileName   = "metadata.json"
 	MaxPinnedScratches = 5
+	storeFileName      = "padz-scratches.json"
 )
 
+// Store implements the padz store interface using nanostore
 type Store struct {
-	mu              sync.Mutex
-	scratches       []Scratch
-	fs              filesystem.FileSystem
-	cfg             *config.Config
-	metadataManager *MetadataManager
-	useNewMetadata  bool // Flag to enable new metadata system
+	store     nanostore.Store
+	fs        filesystem.FileSystem
+	cfg       *config.Config
+	basePath  string
+	filesPath string
 }
 
 func NewStore() (*Store, error) {
-	logger := logging.GetLogger("store")
-	logger.Info().Msg("Initializing new store")
-
-	store, err := NewStoreWithConfig(config.GetConfig())
-	if err != nil {
-		return nil, err
-	}
-
-	// Run migration to ensure project names are relative
-	if err := store.migrateProjectPaths(); err != nil {
-		logger.Warn().Err(err).Msg("Failed to migrate project paths, continuing anyway")
-	}
-
-	return store, nil
+	cfg := config.GetConfig()
+	return NewStoreWithConfig(cfg)
 }
 
 // NewStoreWithScope creates a new store with the specified scope
 func NewStoreWithScope(isGlobal bool) (*Store, error) {
-	logger := logging.GetLogger("store")
-	logger.Info().Bool("is_global", isGlobal).Msg("Initializing store with scope")
 	cfg := config.GetConfig()
 	cfg.IsGlobalScope = isGlobal
-
-	store, err := NewStoreWithConfig(cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	// Run migration to ensure project names are relative
-	if err := store.migrateProjectPaths(); err != nil {
-		logger.Warn().Err(err).Msg("Failed to migrate project paths, continuing anyway")
-	}
-
-	return store, nil
+	return NewStoreWithConfig(cfg)
 }
 
 func NewStoreWithConfig(cfg *config.Config) (*Store, error) {
 	logger := logging.GetLogger("store")
-	logger.Info().Msg("Initializing store with config")
+	logger.Info().Bool("is_global", cfg.IsGlobalScope).Msg("Initializing nanostore")
 
-	store := &Store{
-		fs:  cfg.FileSystem,
-		cfg: cfg,
-	}
-
-	// Initialize metadata manager
-	basePath, err := GetScratchPathWithConfig(cfg)
+	basePath, err := getBasePath(cfg)
 	if err != nil {
 		return nil, err
 	}
-	store.metadataManager = NewMetadataManager(cfg.FileSystem, basePath)
 
-	// Check if we should use new metadata system
-	store.useNewMetadata = store.shouldUseNewMetadata()
-
-	if err := store.load(); err != nil {
-		logger.Error().Err(err).Msg("Failed to load store data")
-		return nil, err
+	// Ensure base directory exists
+	if err := cfg.FileSystem.MkdirAll(basePath, 0755); err != nil {
+		return nil, fmt.Errorf("failed to create base directory: %w", err)
 	}
 
-	logger.Info().Int("scratch_count", len(store.scratches)).Msg("Store initialized successfully")
-	return store, nil
+	// Create files subdirectory
+	filesPath := filepath.Join(basePath, "files")
+	if err := cfg.FileSystem.MkdirAll(filesPath, 0755); err != nil {
+		return nil, fmt.Errorf("failed to create files directory: %w", err)
+	}
+
+	storePath := filepath.Join(basePath, storeFileName)
+
+	// Configure nanostore with padz dimensions
+	config := nanostore.Config{
+		Dimensions: []nanostore.DimensionConfig{
+			{
+				Name:         "activity",
+				Type:         nanostore.Enumerated,
+				Values:       []string{"active", "deleted"},
+				Prefixes:     map[string]string{}, // No prefix for activity
+				DefaultValue: "active",
+			},
+			{
+				Name:         "pinned",
+				Type:         nanostore.Enumerated,
+				Values:       []string{"no", "yes"},
+				Prefixes:     map[string]string{"yes": "p"},
+				DefaultValue: "no",
+			},
+			// Note: Removed project dimension as it causes SimpleID issues
+			// when used as Hierarchical. Project will be stored as regular data.
+		},
+	}
+
+	// Initialize nanostore
+	store, err := nanostore.New(storePath, config)
+	if err != nil {
+		return nil, fmt.Errorf("failed to initialize nanostore: %w", err)
+	}
+
+	return &Store{
+		store:     store,
+		fs:        cfg.FileSystem,
+		cfg:       cfg,
+		basePath:  basePath,
+		filesPath: filesPath,
+	}, nil
 }
 
 func (s *Store) GetScratches() []Scratch {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
 	logger := logging.GetLogger("store")
-	logger.Debug().Int("scratch_count", len(s.scratches)).Msg("Retrieved all scratches")
-	return s.scratches
+
+	// List only active scratches
+	opts := nanostore.ListOptions{
+		Filters: map[string]interface{}{
+			"activity": "active",
+		},
+	}
+
+	docs, err := s.store.List(opts)
+	if err != nil {
+		logger.Error().Err(err).Msg("Failed to query scratches")
+		return []Scratch{}
+	}
+
+	scratches := make([]Scratch, len(docs))
+	for i, doc := range docs {
+		logger.Debug().
+			Str("uuid", doc.UUID).
+			Str("simple_id", doc.SimpleID).
+			Str("title", doc.Title).
+			Msg("Processing document from nanostore")
+		scratches[i] = s.documentToScratch(doc)
+	}
+
+	logger.Debug().Int("count", len(scratches)).Msg("Retrieved scratches")
+	return scratches
 }
 
 func (s *Store) GetPinnedScratches() []Scratch {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
 	logger := logging.GetLogger("store")
-	var pinned []Scratch
-	for _, scratch := range s.scratches {
-		if scratch.IsPinned {
-			pinned = append(pinned, scratch)
-		}
+
+	opts := nanostore.ListOptions{
+		Filters: map[string]interface{}{
+			"activity": "active",
+			"pinned":   "yes",
+		},
 	}
-	logger.Debug().Int("pinned_count", len(pinned)).Msg("Retrieved pinned scratches")
-	return pinned
+
+	docs, err := s.store.List(opts)
+	if err != nil {
+		logger.Error().Err(err).Msg("Failed to query pinned scratches")
+		return []Scratch{}
+	}
+
+	// Sort by pinned time (newest first)
+	scratches := make([]Scratch, len(docs))
+	for i, doc := range docs {
+		scratches[i] = s.documentToScratch(doc)
+	}
+
+	logger.Debug().Int("count", len(scratches)).Msg("Retrieved pinned scratches")
+	return scratches
 }
 
+// SaveScratches performs bulk update
 func (s *Store) SaveScratches(scratches []Scratch) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
 	logger := logging.GetLogger("store")
-	oldCount := len(s.scratches)
-	newCount := len(scratches)
+	logger.Info().Int("count", len(scratches)).Msg("Bulk saving scratches")
 
-	logger.Info().Int("old_count", oldCount).Int("new_count", newCount).Msg("Bulk replacing scratches")
+	// This is a full replacement operation
+	// First, get all existing scratches
+	allOpts := nanostore.ListOptions{}
+	allDocs, err := s.store.List(allOpts)
+	if err != nil {
+		return fmt.Errorf("failed to list existing scratches: %w", err)
+	}
 
-	if s.useNewMetadata {
-		// Rebuild entire metadata system with new scratches
-		// This is expensive but ensures consistency
-		logger.Info().Msg("Rebuilding metadata for bulk save")
-
-		// Create new index
-		index := &Index{
-			Version:   indexVersion,
-			UpdatedAt: time.Now(),
-			Scratches: make(map[string]IndexEntry),
-		}
-
-		// Save all individual metadata files
-		for _, scratch := range scratches {
-			if err := s.metadataManager.SaveScratchMetadata(&scratch); err != nil {
-				logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to save scratch metadata during bulk save")
-				continue
-			}
-
-			// Add to index
-			index.Scratches[scratch.ID] = IndexEntry{
-				Project:   scratch.Project,
-				Title:     scratch.Title,
-				CreatedAt: scratch.CreatedAt,
-			}
-		}
-
-		// Save index
-		if err := s.metadataManager.SaveIndex(index); err != nil {
-			logger.Error().Err(err).Msg("Failed to save index during bulk save")
-			return err
-		}
-
-		// Clean up orphaned metadata files
-		metadataPath := s.metadataManager.GetMetadataPath()
-		if entries, err := s.fs.ReadDir(metadataPath); err == nil {
-			for _, entry := range entries {
-				if strings.HasSuffix(entry.Name(), ".json") {
-					id := strings.TrimSuffix(entry.Name(), ".json")
-					if _, exists := index.Scratches[id]; !exists {
-						// Remove orphaned metadata file
-						_ = s.metadataManager.DeleteScratchMetadata(id)
-					}
-				}
-			}
+	// Delete all existing
+	for _, doc := range allDocs {
+		if err := s.store.Delete(doc.UUID, false); err != nil {
+			logger.Error().Err(err).Str("uuid", doc.UUID).Msg("Failed to delete existing scratch")
 		}
 	}
 
-	s.scratches = scratches
-
-	if !s.useNewMetadata {
-		if err := s.save(); err != nil {
-			logger.Error().Err(err).Int("scratch_count", newCount).Msg("Failed to save scratches after bulk replace")
-			return err
+	// Add all new scratches
+	for _, scratch := range scratches {
+		if err := s.AddScratch(scratch); err != nil {
+			logger.Error().Err(err).Str("title", scratch.Title).Msg("Failed to add scratch in bulk save")
 		}
 	}
 
-	logger.Info().Int("scratch_count", newCount).Msg("Bulk scratches saved successfully")
+	logger.Info().Int("count", len(scratches)).Msg("Bulk save completed")
 	return nil
 }
 
-// shouldUseNewMetadata checks if we should use the new metadata system
-func (s *Store) shouldUseNewMetadata() bool {
-	indexPath := s.metadataManager.GetIndexPath()
-	if _, err := s.fs.Stat(indexPath); err == nil {
-		// Index exists, use new system
-		return true
-	}
-
-	// Check if metadata directory exists with files
-	metadataPath := s.metadataManager.GetMetadataPath()
-	if entries, err := s.fs.ReadDir(metadataPath); err == nil && len(entries) > 0 {
-		// Metadata directory has files, use new system
-		return true
-	}
-
-	return false
-}
-
-func (s *Store) load() error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
-	logger := logging.GetLogger("store")
-
-	if s.useNewMetadata {
-		return s.loadWithNewMetadata()
-	}
-
-	path, err := s.getMetadataPathWithStore()
-	if err != nil {
-		logger.Error().Err(err).Msg("Failed to get metadata path during load")
-		return err
-	}
-
-	logger.Debug().Str("path", path).Msg("Loading store data from metadata file")
-
-	if _, err := s.fs.Stat(path); os.IsNotExist(err) {
-		logger.Info().Str("path", path).Msg("Metadata file does not exist, initializing empty store")
-		s.scratches = []Scratch{}
-		return nil
-	}
-
-	data, err := s.fs.ReadFile(path)
-	if err != nil {
-		logger.Error().Err(err).Str("path", path).Msg("Failed to read metadata file")
-		return err
-	}
-
-	logger.Debug().Int("bytes_read", len(data)).Str("path", path).Msg("Successfully read metadata file")
-
-	if err := json.Unmarshal(data, &s.scratches); err != nil {
-		logger.Error().Err(err).Str("path", path).Int("bytes_read", len(data)).Msg("Failed to unmarshal JSON data")
-		return err
-	}
-
-	logger.Info().Int("scratch_count", len(s.scratches)).Str("path", path).Msg("Store data loaded successfully")
-
-	// Check if we should migrate to new system
-	if !s.useNewMetadata && len(s.scratches) > 0 {
-		logger.Info().Msg("Legacy metadata detected, migrating to new system")
-		if err := s.metadataManager.MigrateFromLegacyMetadata(s.scratches); err != nil {
-			logger.Error().Err(err).Msg("Failed to migrate to new metadata system")
-			// Continue with old system if migration fails
-		} else {
-			s.useNewMetadata = true
-			logger.Info().Msg("Migration to new metadata system completed")
-		}
-	}
-
-	return nil
-}
-
-// loadWithNewMetadata loads scratches using the new metadata system
-func (s *Store) loadWithNewMetadata() error {
-	logger := logging.GetLogger("store")
-	logger.Debug().Msg("Loading store data using new metadata system")
-
-	// Initialize metadata system if needed
-	if err := s.metadataManager.Initialize(); err != nil {
-		return err
-	}
-
-	// Load index
-	index, err := s.metadataManager.LoadIndex()
-	if err != nil {
-		return err
-	}
-
-	// Load individual metadata for each scratch in index
-	s.scratches = make([]Scratch, 0, len(index.Scratches))
-	for id, entry := range index.Scratches {
-		scratch, err := s.metadataManager.LoadScratchMetadata(id)
-		if err != nil {
-			logger.Warn().Err(err).Str("id", id).Msg("Failed to load scratch metadata")
-			continue
-		}
-		if scratch == nil {
-			// Metadata file missing, reconstruct from index
-			scratch = &Scratch{
-				ID:        id,
-				Project:   entry.Project,
-				Title:     entry.Title,
-				CreatedAt: entry.CreatedAt,
-				UpdatedAt: entry.CreatedAt, // Best guess - use creation time
-				Size:      0,               // Unknown without reading file
-				Checksum:  "",              // Unknown without reading file
-			}
-			logger.Warn().
-				Str("id", id).
-				Str("project", entry.Project).
-				Msg("Reconstructed scratch from index - metadata file missing")
-		}
-		s.scratches = append(s.scratches, *scratch)
-	}
-
-	logger.Info().Int("scratch_count", len(s.scratches)).Msg("Store data loaded successfully with new metadata system")
-	return nil
-}
-
-func (s *Store) save() error {
-	logger := logging.GetLogger("store")
-
-	if s.useNewMetadata {
-		return s.saveWithNewMetadata()
-	}
-
-	path, err := s.getMetadataPathWithStore()
-	if err != nil {
-		logger.Error().Err(err).Msg("Failed to get metadata path during save")
-		return err
-	}
-
-	logger.Debug().Str("path", path).Int("scratch_count", len(s.scratches)).Msg("Saving store data to metadata file")
-
-	data, err := json.MarshalIndent(s.scratches, "", "  ")
-	if err != nil {
-		logger.Error().Err(err).Int("scratch_count", len(s.scratches)).Msg("Failed to marshal scratches to JSON")
-		return err
-	}
-
-	logger.Debug().Int("bytes_to_write", len(data)).Str("path", path).Msg("Successfully marshaled data, writing to file")
-
-	if err := s.fs.WriteFile(path, data, 0644); err != nil {
-		logger.Error().Err(err).Str("path", path).Int("bytes_to_write", len(data)).Msg("Failed to write metadata file")
-		return err
-	}
-
-	logger.Debug().Str("path", path).Int("scratch_count", len(s.scratches)).Msg("Store data saved successfully")
-	return nil
-}
-
-// saveWithNewMetadata saves using the new metadata system
-func (s *Store) saveWithNewMetadata() error {
-	logger := logging.GetLogger("store")
-	logger.Debug().Msg("Saving store data using new metadata system")
-
-	// This is now a no-op as individual operations update metadata
-	// The index is updated with each operation
-	logger.Debug().Msg("Save called but using new metadata system (no-op)")
-	return nil
-}
-
+// AddScratch adds a new scratch
 func (s *Store) AddScratch(scratch Scratch) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
 	logger := logging.GetLogger("store")
-	logger.Info().Str("scratch_id", scratch.ID).Str("title", scratch.Title).Msg("Adding new scratch")
+	logger.Info().Str("title", scratch.Title).Msg("Adding scratch")
 
-	if s.useNewMetadata {
-		// Save individual metadata file
-		if err := s.metadataManager.SaveScratchMetadata(&scratch); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to save scratch metadata")
-			return err
+	// Prepare dimensions and data
+	dimensions := map[string]interface{}{
+		"activity": "active",
+		"pinned":   "no",
+	}
+
+	if scratch.IsPinned {
+		dimensions["pinned"] = "yes"
+	}
+
+	// Add project and metadata as data (prefixed with _data.)
+	if scratch.Project != "" {
+		dimensions["_data.project"] = scratch.Project
+	}
+	if scratch.Size > 0 {
+		dimensions["_data.size"] = scratch.Size
+	}
+	if scratch.Checksum != "" {
+		dimensions["_data.checksum"] = scratch.Checksum
+	}
+	if !scratch.PinnedAt.IsZero() {
+		dimensions["_data.pinned_at"] = scratch.PinnedAt
+	}
+
+	// Create the document in nanostore
+	uuid, err := s.store.Add(scratch.Title, dimensions)
+	if err != nil {
+		return fmt.Errorf("failed to create scratch: %w", err)
+	}
+
+	// If there's content, update the document with the body
+	if scratch.Content != "" {
+		updates := nanostore.UpdateRequest{
+			Body: &scratch.Content,
 		}
-
-		// Update index
-		if err := s.metadataManager.UpdateIndexEntry(&scratch); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to update index")
-			return err
+		if err := s.store.Update(uuid, updates); err != nil {
+			logger.Error().Err(err).Msg("Failed to add content to scratch")
+			// Don't fail the whole operation, but log the error
 		}
 	}
 
-	s.scratches = append(s.scratches, scratch)
-
-	if !s.useNewMetadata {
-		if err := s.save(); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to save after adding scratch")
-			return err
-		}
-	}
-
-	logger.Info().Str("scratch_id", scratch.ID).Int("total_count", len(s.scratches)).Msg("Scratch added successfully")
+	logger.Info().Str("uuid", uuid).Str("title", scratch.Title).Msg("Scratch added successfully")
 	return nil
 }
 
+// RemoveScratch soft deletes a scratch
 func (s *Store) RemoveScratch(id string) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
 	logger := logging.GetLogger("store")
-	oldCount := len(s.scratches)
-	logger.Info().Str("scratch_id", id).Int("current_count", oldCount).Msg("Removing scratch")
+	logger.Info().Str("id", id).Msg("Removing scratch")
 
-	if s.useNewMetadata {
-		// Delete individual metadata file
-		if err := s.metadataManager.DeleteScratchMetadata(id); err != nil {
-			logger.Error().Err(err).Str("scratch_id", id).Msg("Failed to delete scratch metadata")
-			return err
-		}
-
-		// Remove from index
-		if err := s.metadataManager.RemoveIndexEntry(id); err != nil {
-			logger.Error().Err(err).Str("scratch_id", id).Msg("Failed to update index")
-			return err
-		}
+	// Find the scratch with this SimpleID by listing all and matching
+	uuid, err := s.resolveSimpleIDToUUID(id)
+	if err != nil {
+		return fmt.Errorf("failed to resolve ID %s: %w", id, err)
 	}
 
-	var newScratches []Scratch
-	found := false
-	for _, scratch := range s.scratches {
-		if scratch.ID != id {
-			newScratches = append(newScratches, scratch)
-		} else {
-			found = true
-			logger.Debug().Str("scratch_id", id).Str("title", scratch.Title).Msg("Found scratch to remove")
-		}
+	// Soft delete by updating activity dimension
+	now := time.Now()
+	updates := nanostore.UpdateRequest{
+		Dimensions: map[string]interface{}{
+			"activity":         "deleted",
+			"_data.deleted_at": now,
+		},
 	}
 
-	if !found {
-		logger.Warn().Str("scratch_id", id).Msg("Scratch not found for removal")
+	logger.Info().Str("uuid", uuid).Interface("updates", updates).Msg("About to update scratch")
+
+	if err := s.store.Update(uuid, updates); err != nil {
+		return fmt.Errorf("failed to remove scratch: %w", err)
 	}
 
-	s.scratches = newScratches
-	newCount := len(s.scratches)
-
-	if !s.useNewMetadata {
-		if err := s.save(); err != nil {
-			logger.Error().Err(err).Str("scratch_id", id).Msg("Failed to save after removing scratch")
-			return err
-		}
-	}
-
-	logger.Info().Str("scratch_id", id).Int("old_count", oldCount).Int("new_count", newCount).Bool("found", found).Msg("Scratch removal completed")
+	logger.Info().Str("id", id).Str("uuid", uuid).Msg("Scratch removed successfully")
 	return nil
 }
 
-func (s *Store) UpdateScratch(scratchToUpdate Scratch) error {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-
+// UpdateScratch updates an existing scratch
+func (s *Store) UpdateScratch(scratch Scratch) error {
 	logger := logging.GetLogger("store")
-	logger.Info().Str("scratch_id", scratchToUpdate.ID).Str("title", scratchToUpdate.Title).Msg("Updating scratch")
+	logger.Info().Str("id", scratch.ID).Str("title", scratch.Title).Msg("Updating scratch")
 
-	if s.useNewMetadata {
-		// Update timestamp
-		scratchToUpdate.UpdatedAt = time.Now()
-
-		// Save individual metadata file
-		if err := s.metadataManager.SaveScratchMetadata(&scratchToUpdate); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratchToUpdate.ID).Msg("Failed to save scratch metadata")
-			return err
-		}
-
-		// Update index
-		if err := s.metadataManager.UpdateIndexEntry(&scratchToUpdate); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratchToUpdate.ID).Msg("Failed to update index")
-			return err
-		}
+	// Resolve SimpleID to UUID for update
+	uuid, err := s.resolveSimpleIDToUUID(scratch.ID)
+	if err != nil {
+		return fmt.Errorf("failed to resolve ID %s: %w", scratch.ID, err)
 	}
 
-	found := false
-	for i, scratch := range s.scratches {
-		if scratch.ID == scratchToUpdate.ID {
-			logger.Debug().Str("scratch_id", scratchToUpdate.ID).Str("old_title", scratch.Title).Str("new_title", scratchToUpdate.Title).Msg("Found scratch to update")
-			s.scratches[i] = scratchToUpdate
-			found = true
-			break
-		}
+	// Prepare update request
+	updates := nanostore.UpdateRequest{
+		Title:      &scratch.Title,
+		Dimensions: map[string]interface{}{
+			// Don't include project - it's not a dimension
+		},
 	}
 
-	if !found {
-		logger.Warn().Str("scratch_id", scratchToUpdate.ID).Msg("Scratch not found for update")
+	// Update body if content is provided
+	if scratch.Content != "" {
+		updates.Body = &scratch.Content
 	}
 
-	if !s.useNewMetadata {
-		if err := s.save(); err != nil {
-			logger.Error().Err(err).Str("scratch_id", scratchToUpdate.ID).Msg("Failed to save after updating scratch")
-			return err
-		}
+	// Update dimensions
+	if scratch.IsPinned {
+		updates.Dimensions["pinned"] = "yes"
+	} else {
+		updates.Dimensions["pinned"] = "no"
 	}
 
-	logger.Info().Str("scratch_id", scratchToUpdate.ID).Bool("found", found).Msg("Scratch update completed")
+	if scratch.IsDeleted {
+		updates.Dimensions["activity"] = "deleted"
+	} else {
+		updates.Dimensions["activity"] = "active"
+	}
+
+	// Update data fields
+	if scratch.Project != "" {
+		updates.Dimensions["_data.project"] = scratch.Project
+	}
+	if scratch.Size > 0 {
+		updates.Dimensions["_data.size"] = scratch.Size
+	}
+	if scratch.Checksum != "" {
+		updates.Dimensions["_data.checksum"] = scratch.Checksum
+	}
+	if !scratch.PinnedAt.IsZero() {
+		updates.Dimensions["_data.pinned_at"] = scratch.PinnedAt
+	}
+	if scratch.IsDeleted && scratch.DeletedAt != nil {
+		updates.Dimensions["_data.deleted_at"] = *scratch.DeletedAt
+	}
+
+	if err := s.store.Update(uuid, updates); err != nil {
+		return fmt.Errorf("failed to update scratch: %w", err)
+	}
+
+	logger.Info().Str("id", scratch.ID).Str("uuid", uuid).Msg("Scratch updated successfully")
 	return nil
 }
 
+// Close closes the store
+func (s *Store) Close() error {
+	return s.store.Close()
+}
+
+// Search searches for scratches matching query
+func (s *Store) Search(query string) []Scratch {
+	logger := logging.GetLogger("store")
+
+	opts := nanostore.ListOptions{
+		FilterBySearch: query,
+		Filters: map[string]interface{}{
+			"activity": "active",
+		},
+	}
+
+	docs, err := s.store.List(opts)
+	if err != nil {
+		logger.Error().Err(err).Str("query", query).Msg("Failed to search scratches")
+		return []Scratch{}
+	}
+
+	scratches := make([]Scratch, len(docs))
+	for i, doc := range docs {
+		scratches[i] = s.documentToScratch(doc)
+	}
+
+	return scratches
+}
+
+// documentToScratch converts a nanostore Document to a Scratch
+func (s *Store) documentToScratch(doc nanostore.Document) Scratch {
+	scratch := Scratch{
+		ID:        doc.SimpleID,
+		Title:     doc.Title,
+		Content:   doc.Body, // Content is stored in nanostore's Body field
+		CreatedAt: doc.CreatedAt,
+		UpdatedAt: doc.UpdatedAt,
+		IsPinned:  s.getDocumentDimension(doc, "pinned") == "yes",
+		IsDeleted: s.getDocumentDimension(doc, "activity") == "deleted",
+	}
+
+	// Extract data fields from dimensions (prefixed with _data.)
+	if project, ok := doc.Dimensions["_data.project"].(string); ok {
+		scratch.Project = project
+	}
+	if size, ok := doc.Dimensions["_data.size"].(float64); ok {
+		scratch.Size = int64(size)
+	}
+	if checksum, ok := doc.Dimensions["_data.checksum"].(string); ok {
+		scratch.Checksum = checksum
+	}
+	if pinnedAt, ok := doc.Dimensions["_data.pinned_at"].(time.Time); ok {
+		scratch.PinnedAt = pinnedAt
+	}
+	// Handle deleted_at
+	if scratch.IsDeleted {
+		if deletedAt, ok := doc.Dimensions["_data.deleted_at"].(time.Time); ok {
+			scratch.DeletedAt = &deletedAt
+		}
+	}
+
+	return scratch
+}
+
+// getDocumentDimension safely extracts a dimension value from document
+func (s *Store) getDocumentDimension(doc nanostore.Document, dimension string) string {
+	if val, ok := doc.Dimensions[dimension].(string); ok {
+		return val
+	}
+	return ""
+}
+
+// Atomic operations (delegate to non-atomic versions as nanostore handles locking)
+func (s *Store) AddScratchAtomic(scratch Scratch) error {
+	return s.AddScratch(scratch)
+}
+
+func (s *Store) RemoveScratchAtomic(id string) error {
+	return s.RemoveScratch(id)
+}
+
+func (s *Store) UpdateScratchAtomic(scratch Scratch) error {
+	return s.UpdateScratch(scratch)
+}
+
+func (s *Store) SaveScratchesAtomic(scratches []Scratch) error {
+	return s.SaveScratches(scratches)
+}
+
+// GetScratchPath returns the file path for a scratch's content
+func (s *Store) GetScratchPath(id string) (string, error) {
+	// Resolve to UUID
+	uuid, err := s.store.ResolveUUID(id)
+	if err != nil {
+		return "", err
+	}
+	return filepath.Join(s.filesPath, uuid), nil
+}
+
+// RebuildIndex rebuilds the master index (no-op for nanostore)
+func (s *Store) RebuildIndex() error {
+	// nanostore handles indexing automatically
+	return nil
+}
+
+// RunDiscoveryBeforeCommand discovers orphaned files (no-op for nanostore)
+func (s *Store) RunDiscoveryBeforeCommand() error {
+	// nanostore stores content in the body field, not as separate files
+	// so there's no need for discovery of orphaned files
+	return nil
+}
+
+// resolveSimpleIDToUUID finds the UUID for a given SimpleID by listing all documents
+func (s *Store) resolveSimpleIDToUUID(simpleID string) (string, error) {
+	// List all documents (including deleted ones) to find the one with this SimpleID
+	allOpts := nanostore.ListOptions{}
+	docs, err := s.store.List(allOpts)
+	if err != nil {
+		return "", fmt.Errorf("failed to list documents: %w", err)
+	}
+
+	for _, doc := range docs {
+		if doc.SimpleID == simpleID {
+			return doc.UUID, nil
+		}
+	}
+
+	return "", fmt.Errorf("scratch not found: %s", simpleID)
+}
+
+// Helper functions
+
+func getBasePath(cfg *config.Config) (string, error) {
+	if cfg.DataPath != "" {
+		return filepath.Join(cfg.DataPath, "scratch"), nil
+	}
+
+	if cfg.IsGlobalScope {
+		// Use XDG for global scope
+		homeDir, err := os.UserHomeDir()
+		if err != nil {
+			return "", fmt.Errorf("failed to get home directory: %w", err)
+		}
+		return filepath.Join(homeDir, ".local", "share", "padz", "scratch"), nil
+	}
+
+	// Use local .padz directory for project scope
+	projectRoot, err := project.GetProjectRoot(".")
+	if err != nil || projectRoot == "" {
+		// Not in a project, fall back to global
+		homeDir, err := os.UserHomeDir()
+		if err != nil {
+			return "", fmt.Errorf("failed to get home directory: %w", err)
+		}
+		return filepath.Join(homeDir, ".local", "share", "padz", "scratch"), nil
+	}
+
+	return filepath.Join(projectRoot, ".padz"), nil
+}
+
+// Legacy compatibility functions
 func GetScratchPath() (string, error) {
 	cfg := config.GetConfig()
-	return GetScratchPathWithConfig(cfg)
+	return getBasePath(cfg)
 }
 
 func GetScratchPathWithConfig(cfg *config.Config) (string, error) {
-	logger := logging.GetLogger("store.path")
-	var path string
-	var err error
-
-	if cfg.DataPath != "" {
-		// Use configured path for testing
-		path = cfg.FileSystem.Join(cfg.DataPath, dataDirName)
-		logger.Debug().Str("path", path).Msg("Using configured data path")
-	} else if cfg.IsGlobalScope {
-		// Use XDG for global scope
-		path, err = xdg.DataFile(dataDirName)
-		if err != nil {
-			logger.Error().Err(err).Msg("Failed to get XDG data file path")
-			return "", err
-		}
-		logger.Debug().Str("path", path).Msg("Using XDG data path for global scope")
-	} else {
-		// Use local .padz directory for project scope
-		currentDir, err := os.Getwd()
-		if err != nil {
-			logger.Error().Err(err).Msg("Failed to get current working directory")
-			return "", err
-		}
-
-		projectRoot, err := project.GetProjectRoot(currentDir)
-		if err != nil {
-			logger.Error().Err(err).Msg("Failed to get project root")
-			return "", err
-		}
-
-		if projectRoot == "" {
-			// Not in a project, fall back to XDG
-			path, err = xdg.DataFile(dataDirName)
-			if err != nil {
-				logger.Error().Err(err).Msg("Failed to get XDG data file path")
-				return "", err
-			}
-			logger.Debug().Str("path", path).Msg("Not in project, using XDG data path")
-		} else {
-			// In a project, use .padz directory
-			path = filepath.Join(projectRoot, ".padz")
-			logger.Debug().Str("path", path).Str("project_root", projectRoot).Msg("Using local .padz directory")
-		}
-	}
-
-	logger.Debug().Str("path", path).Msg("Creating scratch directory if needed")
-	if err := cfg.FileSystem.MkdirAll(path, 0755); err != nil {
-		logger.Error().Err(err).Str("path", path).Msg("Failed to create scratch directory")
-		return "", err
-	}
-
-	logger.Debug().Str("path", path).Msg("Scratch path resolved successfully")
-	return path, nil
+	return getBasePath(cfg)
 }
 
 func GetScratchFilePath(id string) (string, error) {
@@ -555,331 +505,11 @@ func GetScratchFilePath(id string) (string, error) {
 }
 
 func GetScratchFilePathWithConfig(id string, cfg *config.Config) (string, error) {
-	logger := logging.GetLogger("store.path")
-
-	path, err := GetScratchPathWithConfig(cfg)
+	store, err := NewStoreWithConfig(cfg)
 	if err != nil {
-		logger.Error().Err(err).Str("scratch_id", id).Msg("Failed to get base scratch path")
 		return "", err
 	}
+	defer func() { _ = store.Close() }()
 
-	// Check if new metadata system is in use
-	metadataManager := NewMetadataManager(cfg.FileSystem, path)
-	filesPath := metadataManager.GetFilesPath()
-	if _, err := cfg.FileSystem.Stat(filesPath); err == nil {
-		// New structure exists, use it
-		filePath := cfg.FileSystem.Join(filesPath, id)
-		logger.Debug().Str("scratch_id", id).Str("file_path", filePath).Msg("Resolved scratch file path (new structure)")
-		return filePath, nil
-	}
-
-	// Fall back to old structure
-	filePath := cfg.FileSystem.Join(path, id)
-	logger.Debug().Str("scratch_id", id).Str("file_path", filePath).Msg("Resolved scratch file path (legacy structure)")
-	return filePath, nil
-}
-
-func getMetadataPathWithConfig(cfg *config.Config) (string, error) {
-	logger := logging.GetLogger("store.path")
-
-	path, err := GetScratchPathWithConfig(cfg)
-	if err != nil {
-		logger.Error().Err(err).Msg("Failed to get base scratch path for metadata")
-		return "", err
-	}
-
-	metadataPath := cfg.FileSystem.Join(path, metadataFileName)
-	logger.Debug().Str("metadata_path", metadataPath).Msg("Resolved metadata file path")
-	return metadataPath, nil
-}
-
-func (s *Store) getMetadataPathWithStore() (string, error) {
-	return getMetadataPathWithConfig(s.cfg)
-}
-
-// RebuildIndex rebuilds the master index from individual metadata files
-func (s *Store) RebuildIndex() error {
-	if !s.useNewMetadata {
-		return fmt.Errorf("new metadata system not in use")
-	}
-
-	return s.metadataManager.RebuildIndex()
-}
-
-// withFileLock performs an operation with file-based locking to prevent concurrent metadata corruption
-func (s *Store) withFileLock(operation func() error) error {
-	logger := logging.GetLogger("store")
-
-	var lockPath string
-	var err error
-
-	if s.useNewMetadata {
-		// Lock on index file for new system
-		lockPath = s.metadataManager.GetIndexPath()
-	} else {
-		// Lock on metadata.json for old system
-		lockPath, err = s.getMetadataPathWithStore()
-		if err != nil {
-			return err
-		}
-	}
-
-	lock := filelock.New(lockPath)
-
-	// Try to acquire lock with 5 second timeout
-	if err := lock.Lock(5 * time.Second); err != nil {
-		logger.Error().Err(err).Msg("Failed to acquire file lock")
-		return err
-	}
-	defer func() {
-		if unlockErr := lock.Unlock(); unlockErr != nil {
-			logger.Error().Err(unlockErr).Msg("Failed to release file lock")
-		}
-	}()
-
-	// Reload the latest data while holding the lock
-	if err := s.load(); err != nil {
-		return err
-	}
-
-	// Perform the operation
-	if err := operation(); err != nil {
-		return err
-	}
-
-	// Save the changes
-	return s.save()
-}
-
-// AddScratchAtomic adds a scratch with file locking to prevent concurrent conflicts
-func (s *Store) AddScratchAtomic(scratch Scratch) error {
-	// If using memory filesystem (for tests), fall back to non-atomic
-	if _, ok := s.fs.(*filesystem.MemoryFileSystem); ok {
-		return s.AddScratch(scratch)
-	}
-
-	return s.withFileLock(func() error {
-		logger := logging.GetLogger("store")
-
-		// Check for duplicates
-		for _, existing := range s.scratches {
-			if existing.ID == scratch.ID {
-				logger.Warn().Str("scratch_id", scratch.ID).Msg("Scratch with this ID already exists, skipping add")
-				return nil
-			}
-		}
-
-		logger.Info().Str("scratch_id", scratch.ID).Str("title", scratch.Title).Msg("Adding new scratch atomically")
-
-		if s.useNewMetadata {
-			// Save individual metadata file (no lock needed for individual file)
-			if err := s.metadataManager.SaveScratchMetadata(&scratch); err != nil {
-				return err
-			}
-
-			// Update index (already within lock)
-			if err := s.metadataManager.UpdateIndexEntry(&scratch); err != nil {
-				return err
-			}
-		}
-
-		s.scratches = append(s.scratches, scratch)
-		return nil
-	})
-}
-
-// RemoveScratchAtomic removes a scratch with file locking
-func (s *Store) RemoveScratchAtomic(id string) error {
-	// If using memory filesystem (for tests), fall back to non-atomic
-	if _, ok := s.fs.(*filesystem.MemoryFileSystem); ok {
-		return s.RemoveScratch(id)
-	}
-
-	return s.withFileLock(func() error {
-		logger := logging.GetLogger("store")
-
-		var newScratches []Scratch
-		found := false
-
-		for _, scratch := range s.scratches {
-			if scratch.ID != id {
-				newScratches = append(newScratches, scratch)
-			} else {
-				found = true
-				logger.Debug().Str("scratch_id", id).Str("title", scratch.Title).Msg("Found scratch to remove")
-			}
-		}
-
-		if !found {
-			logger.Warn().Str("scratch_id", id).Msg("Scratch not found for removal")
-		}
-
-		if s.useNewMetadata && found {
-			// Delete individual metadata file
-			if err := s.metadataManager.DeleteScratchMetadata(id); err != nil {
-				return err
-			}
-
-			// Remove from index
-			if err := s.metadataManager.RemoveIndexEntry(id); err != nil {
-				return err
-			}
-		}
-
-		s.scratches = newScratches
-		return nil
-	})
-}
-
-// UpdateScratchAtomic updates a scratch with file locking
-func (s *Store) UpdateScratchAtomic(scratchToUpdate Scratch) error {
-	// If using memory filesystem (for tests), fall back to non-atomic
-	if _, ok := s.fs.(*filesystem.MemoryFileSystem); ok {
-		return s.UpdateScratch(scratchToUpdate)
-	}
-
-	return s.withFileLock(func() error {
-		logger := logging.GetLogger("store")
-
-		found := false
-		for i, scratch := range s.scratches {
-			if scratch.ID == scratchToUpdate.ID {
-				logger.Debug().Str("scratch_id", scratchToUpdate.ID).Msg("Found scratch to update")
-				s.scratches[i] = scratchToUpdate
-				found = true
-				break
-			}
-		}
-
-		if !found {
-			logger.Warn().Str("scratch_id", scratchToUpdate.ID).Msg("Scratch not found for update")
-		}
-
-		if s.useNewMetadata && found {
-			// Update timestamp
-			scratchToUpdate.UpdatedAt = time.Now()
-
-			// Save individual metadata file
-			if err := s.metadataManager.SaveScratchMetadata(&scratchToUpdate); err != nil {
-				return err
-			}
-
-			// Update index
-			if err := s.metadataManager.UpdateIndexEntry(&scratchToUpdate); err != nil {
-				return err
-			}
-		}
-
-		return nil
-	})
-}
-
-// SaveScratchesAtomic performs bulk update with file locking
-func (s *Store) SaveScratchesAtomic(scratches []Scratch) error {
-	// If using memory filesystem (for tests), fall back to non-atomic
-	if _, ok := s.fs.(*filesystem.MemoryFileSystem); ok {
-		return s.SaveScratches(scratches)
-	}
-
-	return s.withFileLock(func() error {
-		logger := logging.GetLogger("store")
-		logger.Info().Int("old_count", len(s.scratches)).Int("new_count", len(scratches)).Msg("Bulk replacing scratches atomically")
-
-		if s.useNewMetadata {
-			// Rebuild metadata within lock
-			// Create new index
-			index := &Index{
-				Version:   indexVersion,
-				UpdatedAt: time.Now(),
-				Scratches: make(map[string]IndexEntry),
-			}
-
-			// Save all individual metadata files
-			for _, scratch := range scratches {
-				if err := s.metadataManager.SaveScratchMetadata(&scratch); err != nil {
-					logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to save scratch metadata during bulk save")
-					continue
-				}
-
-				// Add to index
-				index.Scratches[scratch.ID] = IndexEntry{
-					Project:   scratch.Project,
-					Title:     scratch.Title,
-					CreatedAt: scratch.CreatedAt,
-				}
-			}
-
-			// Save index
-			if err := s.metadataManager.SaveIndex(index); err != nil {
-				return err
-			}
-		}
-
-		s.scratches = scratches
-		return nil
-	})
-}
-
-// migrateProjectPaths ensures project paths are stored as relative names, not absolute paths
-func (s *Store) migrateProjectPaths() error {
-	logger := logging.GetLogger("store")
-
-	// Only run migration if we're in a local project context
-	if s.cfg.IsGlobalScope {
-		return nil
-	}
-
-	// Get current project name
-	currentDir, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-
-	projectRoot, err := project.GetProjectRoot(currentDir)
-	if err != nil || projectRoot == "" {
-		// Not in a project, nothing to migrate
-		return nil
-	}
-
-	currentProjectName := filepath.Base(projectRoot)
-	modified := false
-
-	// Check each scratch for absolute paths in the project field
-	for i, scratch := range s.scratches {
-		// If project field contains a path separator, it might be an absolute path
-		if strings.Contains(scratch.Project, string(filepath.Separator)) {
-			// Extract just the base name
-			newProjectName := filepath.Base(scratch.Project)
-
-			// Only update if it matches our current project
-			if newProjectName == currentProjectName || scratch.Project == projectRoot {
-				logger.Info().
-					Str("scratch_id", scratch.ID).
-					Str("old_project", scratch.Project).
-					Str("new_project", currentProjectName).
-					Msg("Migrating absolute project path to relative")
-
-				s.scratches[i].Project = currentProjectName
-				modified = true
-
-				// Update metadata if using new system
-				if s.useNewMetadata {
-					s.scratches[i].UpdatedAt = time.Now()
-					if err := s.metadataManager.SaveScratchMetadata(&s.scratches[i]); err != nil {
-						logger.Error().Err(err).Str("scratch_id", scratch.ID).Msg("Failed to update scratch metadata during migration")
-					}
-				}
-			}
-		}
-	}
-
-	// Save changes if any modifications were made
-	if modified {
-		logger.Info().Msg("Project path migration completed, saving changes")
-		if !s.useNewMetadata {
-			return s.save()
-		}
-	}
-
-	return nil
+	return store.GetScratchPath(id)
 }
